requirements.txt
torch
torchvision
matplotlib
tkinter
pillow
numpy